Maintain Original LatLng Projection In Input Point Properties (Lat, Lon)
- The MMW frontend uses the returned feature's properties to show the original point (as opposed to the snapped one). We were returning it in its projected EPSG: 5070 coordinates, not the original LatLng projection
